Add range metadata, simplify conditional folding
